Command.Build is compiling
This is now compiling and looks OK.  A few points:

(0) Printing of SyntacticMarkers remains to be figured out.

(1) I wonder whether wybs.lang.Build.Task requires a type parameter.

(2) Lots of other commands are *not* compiling.  This is challenging
because they require information currently unavailable (e.g. clean needs
to know what is a build artifact).